Add a description of all commands available
/cc @puzza007 @rpt @IbnFirnas @aerosol @gornikm

Here's a description on how to use onan. Take a look over and maybe even
test it out if you want. You'd need to have the onan server running.

The only note is that the bootstrap command is only partially
implemented.

Let me know also if you have any doubts about the way this will work